move couch from fixtures into mocks
